+++ ignite/cmd/scaffold_chain.go
@@ -3,6 +3,8 @@ package ignitecmd
 import (
 	"github.com/spf13/cobra"
 
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
 	"github.com/ignite/cli/v29/ignite/config/chain/defaults"
 	"github.com/ignite/cli/v29/ignite/pkg/cliui"
 	"github.com/ignite/cli/v29/ignite/pkg/env"
@@ -17,6 +19,7 @@ const (
 	flagMinimal         = "minimal"
 	flagNoDefaultModule = "no-module"
 	flagSkipGit         = "skip-git"
+	flagDefaultDenom    = "default-denom"
 
 	tplScaffoldChainSuccess = `
 ⭐️ Successfully created a new blockchain '%[1]v'.
@@ -87,6 +90,7 @@ about Cosmos SDK on https://docs.cosmos.network
 	flagSetClearCache(c)
 	c.Flags().AddFlagSet(flagSetAccountPrefixes())
 	c.Flags().AddFlagSet(flagSetCoinType())
+	c.Flags().String(flagDefaultDenom, sdk.DefaultBondDenom, "default staking denom")
 	c.Flags().StringP(flagPath, "p", "", "create a project in a specific path")
 	c.Flags().Bool(flagNoDefaultModule, false, "create a project without a default module")
 	c.Flags().StringSlice(flagParams, []string{}, "add default module parameters")
@@ -119,7 +123,12 @@ func scaffoldChainHandler(cmd *cobra.Command, args []string) error {
 		moduleConfigs, _   = cmd.Flags().GetStringSlice(flagModuleConfigs)
 		skipProto, _       = cmd.Flags().GetBool(flagSkipProto)
 		protoDir, _        = cmd.Flags().GetString(flagProtoDir)
+		defaultDenom, _    = cmd.Flags().GetString(flagDefaultDenom)
 	)
+
+	if cmd.Flags().Changed(flagDefaultDenom) && len(defaultDenom) <= 2 {
+		return errors.New("default denom must be at least 3 characters and maximum 128 characters")
+	}
 
 	if noDefaultModule {
 		if len(params) > 0 {
@@ -143,6 +152,7 @@ func scaffoldChainHandler(cmd *cobra.Command, args []string) error {
 		name,
 		addressPrefix,
 		coinType,
+		defaultDenom,
 		protoDir,
 		noDefaultModule,
 		minimal,