+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Features
+
+- [#4683](https://github.com/ignite/cli/pull/4683) Allow to change default denom via flag.
+
 ## [`v29.0.0-rc.1`](https://github.com/ignite/cli/releases/tag/v29.0.0-rc.1)
 
 ### Features